+++ R/boottest_felm.R
@@ -27,8 +27,9 @@
 #'        intervals by p-value inversion. If FALSE, only the p-value is returned.
 #' @param seed An integer. Allows the user to set a random seed. If you want to set a "global" seed, set it via `dqrng::dqset.seed()`. For Mammen weights, you have to use `set.seed()` instead. 
 #' @param R Hypothesis Vector giving linear combinations of coefficients. Must be either NULL or a vector of the same length as `param`. If NULL, a vector of ones of length param.
-#' @param beta0 A numeric. Shifts the null hypothesis 
-#'        H0: param = beta0 vs H1: param != beta0
+#' @param r A numeric. Shifts the null hypothesis 
+#'        H0: param = r vs H1: param != r
+#' @param beta0 Superseded function argument, replaced by function argument 'r'        
 #' @param type character or function. The character string specifies the type
 #'        of boostrap to use: One of "rademacher", "mammen", "norm"
 #'        and "webb". Alternatively, type can be a function(n) for drawing 
@@ -87,8 +88,8 @@
 #' \item{sign_level}{Significance level used in boottest.}
 #' \item{type}{Distribution of the bootstrap weights.}
 #' \item{impose_null}{Whether the null was imposed on the bootstrap dgp or not.}
-#' \item{R}{The vector "R" in the null hypothesis of interest Rbeta = beta0.}
-#' \item{beta0}{The scalar "beta0" in the null hypothesis of interest Rbeta = beta0.}
+#' \item{R}{The vector "R" in the null hypothesis of interest Rbeta = r.}
+#' \item{r}{The scalar "r" in the null hypothesis of interest Rbeta = r.}
 #' \item{point_estimate}{R'beta. A scalar: the constraints vector times the regression coefficients.}
 #' \item{p_test_vals}{All p-values calculated while calculating the confidence
 #'      interval.}
@@ -153,14 +154,14 @@
 #'                   fe = "Q1_immigration", 
 #'                   sign_level = 0.2, 
 #'                   seed = 8,
-#'                   beta0 = 2)
+#'                   r = 2)
 #' # test treatment + ideology1 = 2                   
 #' boot5 <- boottest(felm_fit, 
 #'                   B = 9999, 
 #'                   clustid = c("group_id1", "group_id2"),
 #'                   param = c("treatment", "ideology1"),
 #'                   R = c(1, 1), 
-#'                   beta0 = 2)
+#'                   r = 2)
 #' summary(boot1)
 #' plot(boot1)
 #' }
@@ -177,7 +178,8 @@ boottest.felm <- function(object,
                           conf_int = TRUE,
                           seed = NULL,
                           R = NULL,
-                          beta0 = 0,
+                          r = 0,
+                          beta0 = r, 
                           sign_level = 0.05,
                           type = "rademacher",
                           impose_null = TRUE,
@@ -214,6 +216,7 @@ boottest.felm <- function(object,
   check_arg(conf_int, "logical scalar")
   check_arg(seed, "scalar integer | NULL")
   check_arg(R, "NULL| scalar numeric | numeric vector")
+  check_arg(r, "numeric scalar | NULL")
   check_arg(beta0, "numeric scalar | NULL")
   check_arg(fe, "character scalar | NULL")
   check_arg(bootcluster, "character vector")
@@ -225,6 +228,11 @@ boottest.felm <- function(object,
   check_arg(floattype, "charin(Float32, Float64)")
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
+  
+  if(missing(r) & !missing(beta0)){
+    warning("Note that the 'beta0' function argument is superseded by a new argument, 'r'. Please specify your hypothesis via the new function argument instead of using 'beta0'.")
+  }
+  
   
   if(boot_algo == "R-lean"){
     if(!is.null(fe)){
@@ -395,7 +403,7 @@ boottest.felm <- function(object,
                         boot_iter = B,
                         point_estimate = point_estimate,
                         impose_null = impose_null,
-                        beta0 = beta0,
+                        r = r,
                         sign_level = sign_level,
                         param = param,
                         # seed = seed,
@@ -410,7 +418,7 @@ boottest.felm <- function(object,
                         boot_iter = B,
                         point_estimate = point_estimate,
                         impose_null = impose_null,
-                        beta0 = beta0,
+                        r = r,
                         sign_level = sign_level,
                         param = param,
                         # seed = seed,
@@ -438,7 +446,7 @@ boottest.felm <- function(object,
         # sign, abs for security
         se_guess <- abs(point_estimate / res$t_stat)
       } else if(impose_null == FALSE){
-        se_guess <- abs((point_estimate - beta0) / res$t_stat)
+        se_guess <- abs((point_estimate - r) / res$t_stat)
       }
 
 
@@ -483,7 +491,7 @@ boottest.felm <- function(object,
       type = type,
       impose_null = impose_null,
       R = R,
-      beta0 = beta0, 
+      r = r, 
       boot_algo = boot_algo, 
       nthreads = nthreads
       )
@@ -539,7 +547,7 @@ boottest.felm <- function(object,
     } else {
       R <- matrix(preprocess$R, 1, length(preprocess$R))
     }
-    r <- beta0
+    r <- r
     reps <- as.integer(B) # WildBootTests.jl demands integer
     
     # Order the columns of `clustid` this way:
@@ -676,7 +684,7 @@ boottest.felm <- function(object,
       type = type,
       impose_null = impose_null,
       R = R,
-      beta0 = beta0,
+      r = r,
       plotpoints = plotpoints, 
       # boot_algo returns NULL if not set via global variable
       boot_algo = "WildBootTests.jl"
@@ -714,8 +722,8 @@ boottest.felm <- function(object,
 #' @param seed An integer. Controls the random number generation, which is handled via the `StableRNG()` function from the `StableRNGs` Julia package.
 #' @param R Hypothesis Vector or Matrix giving linear combinations of coefficients. Must be either a vector of length k or a matrix of dimension q x k, where q is the number
 #'        of joint hypotheses and k the number of estimated coefficients.
-#' @param beta0 A vector of length q, where q is the number of tested hypotheses. Shifts the null hypothesis
-#'        H0: param = beta0 vs H1: param != beta0. If not provided, a vector of zeros of length q.
+#' @param r A vector of length q, where q is the number of tested hypotheses. Shifts the null hypothesis
+#'        H0: param = r vs H1: param != r. If not provided, a vector of zeros of length q.
 #' @param type character or function. The character string specifies the type
 #'        of boostrap to use: One of "rademacher", "mammen", "norm", "gamma"
 #'        and "webb". Alternatively, type can be a function(n) for drawing
@@ -797,7 +805,8 @@ waldboottest.felm <- function(object,
                                 clustid,
                                 B,
                                 R,
-                                beta0 = rep(0,nrow(R)),
+                                r = rep(0,nrow(R)),
+                                beta0 = r,
                                 bootcluster = "max",
                                 fe = NULL, 
                                 seed = NULL,
@@ -833,7 +842,7 @@ waldboottest.felm <- function(object,
   check_arg(p_val_type, 'charin(two-tailed, equal-tailed,>, <)')
   
   check_arg(seed, "scalar integer | NULL")
-  check_arg(beta0, "numeric scalar | NULL")
+  check_arg(r, "numeric scalar | NULL")
   check_arg(fe, "character scalar | NULL")
   check_arg(bootcluster, "character vector")
   check_arg(tol, "numeric scalar GT{0}")
@@ -844,7 +853,6 @@ waldboottest.felm <- function(object,
   check_arg(maxmatsize, "scalar integer | NULL")
   check_arg(bootstrapc, "scalar logical")
   
-
   if(is.null(seed)){
     internal_seed <- get_seed()
   } else {
@@ -937,7 +945,7 @@ waldboottest.felm <- function(object,
     } else {
       R <- matrix(preprocess$R, 1, length(preprocess$R))
     }
-    r <- beta0
+    r <- r
     reps <- as.integer(B) # WildBootTests.jl demands integer
     
     # Order the columns of `clustid` this way:
@@ -1068,7 +1076,7 @@ waldboottest.felm <- function(object,
       type = type,
       impose_null = impose_null,
       R = R,
-      beta0 = beta0,
+      r = r,
       plotpoints = plotpoints
     )